# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: OpportuneX API
  description: |
    OpportuneX is an AI-powered platform that helps students from Tier 2 and Tier 3 cities in India discover and access hackathons, internships, and workshops through natural language interactions.
    
    ## Features
    - **Smart Search**: Natural language search for opportunities with advanced filtering
    - **Voice Search**: Multi-language voice commands (English and Hindi)
    - **AI Instructor**: Personalized preparation roadmaps and guidance
    - **Smart Notifications**: Personalized opportunity alerts via email, SMS, and in-app
    
    ## Authentication
    Most endpoints require authentication using JWT tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    - General API endpoints: 100 requests per 15 minutes
    - Search endpoints: 50 requests per 15 minutes
    - Authentication endpoints: 10 requests per 15 minutes
    
    ## Error Handling
    The API uses standard HTTP status codes and returns error details in JSON format:
    ```json
    {
      "error": {
        "code": "VALIDATION_ERROR",
        "message": "Invalid request parameters",
        "details": ["Field 'query' is required"]
      }
    }
    ```
  version: 1.0.0
  contact:
    name: OpportuneX API Support
    email: api-support@opportunex.com
    url: https://opportunex.com/support
  license:
    name: MIT
...
servers:
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: "student@example.com"
              password: "SecurePassword123!"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimit'
...
  /user/profile:
    get:
      tags:
        - User Profile
      summary: Get user profile
      description: Retrieve the authenticated user's profile information
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags:
        - User Profile
      summary: Update user profile
      description: Update the authenticated user's profile information
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdate'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /opportunities:
    get:
      tags:
        - Opportunities
      summary: Get opportunities list
      description: Retrieve a paginated list of opportunities with optional filtering
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: ["hackathon", "internship", "workshop"]
          description: Filter by opportunity type
        - name: skills
          in: query
          schema:
            type: array
            items:
              type: string
          style: form
          explode: false
          description: Filter by required skills
        - name: location
          in: query
          schema:
            type: string
          description: Filter by location
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
          description: Number of items per page
      responses:
        '200':
          description: Opportunities list
          content:
            application/json:
              schema:
                type: object
                properties:
                  opportunities:
                    type: array
                    items:
                      $ref: '#/components/schemas/Opportunity'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                  totalCount:
                    type: integer
...
  /notifications:
    get:
      tags:
        - Notifications
      summary: Get user notifications
      description: Retrieve user's notifications with pagination
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: ["unread", "read", "all"]
            default: "all"
          description: Filter by notification status
        - name: type
          in: query
          schema:
            type: string
            enum: ["opportunity_match", "deadline_reminder", "roadmap_update"]
          description: Filter by notification type
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: Notifications list
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserProfile'
        tokens:
          type: object
          properties:
            accessToken:
              type: string
              description: JWT access token
            refreshToken:
              type: string
              description: Refresh token
            expiresIn:
              type: integer
              description: Token expiration time in seconds

    # Search Schemas
...
    SearchFilters:
      type: object
      properties:
        skills:
          type: array
          items:
            type: string
          description: Required technical skills
        organizerType:
          type: string
          enum: ["corporate", "startup", "government", "academic"]
          description: Type of organizing entity
        mode:
          type: string
          enum: ["online", "offline", "hybrid"]
          description: Event mode
        location:
          type: string
          description: Geographic location
        type:
          type: string
          enum: ["hackathon", "internship", "workshop"]
          description: Opportunity type
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          description: User ID
        email:
          type: string
          format: email
        name:
          type: string
        phone:
          type: string
        location:
          $ref: '#/components/schemas/Location'
        academic:
          $ref: '#/components/schemas/Academic'
        skills:
          $ref: '#/components/schemas/Skills'
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        searchHistory:
          type: array
          items:
            $ref: '#/components/schemas/SearchQuery'
        favoriteOpportunities:
          type: array
          items:
            type: string
          description: Array of favorite opportunity IDs
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserProfileUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        location:
          $ref: '#/components/schemas/Location'
        academic:
          $ref: '#/components/schemas/Academic'
        skills:
          $ref: '#/components/schemas/Skills'
    Location:
      type: object
      required:
        - city
        - state
        - tier
      properties:
        city:
          type: string
          description: City name
        state:
          type: string
          description: State name
        tier:
          type: integer
          enum: [2, 3]
          description: City tier (2 or 3)
    Academic:
      type: object
      required:
        - institution
        - degree
        - year
      properties:
        institution:
          type: string
          description: Educational institution name
        degree:
          type: string
          description: Degree program
        year:
          type: integer
          minimum: 1
          maximum: 6
          description: Current year of study
        cgpa:
          type: number
          minimum: 0
          maximum: 10
          description: Current CGPA (optional)
    Skills:
      type: object
      properties:
        technical:
          type: array
          items:
            type: string
          description: Technical skills
        domains:
          type: array
          items:
            type: string
          description: Domain expertise areas
        proficiencyLevel:
          type: string
          enum: ["beginner", "intermediate", "advanced"]
          description: Overall proficiency level
    UserPreferences:
      type: object
      properties:
        opportunityTypes:
          type: array
          items:
            type: string
            enum: ["hackathon", "internship", "workshop"]
          description: Preferred opportunity types
        preferredMode:
          type: string
          enum: ["online", "offline", "hybrid", "any"]
          description: Preferred event mode
        maxDistance:
          type: integer
          minimum: 0
          description: Maximum distance for offline events (km)
        notifications:
          $ref: '#/components/schemas/NotificationPreferences'
    NotificationPreferences:
      type: object
      properties:
        email:
          type: boolean
          description: Enable email notifications
        sms:
          type: boolean
          description: Enable SMS notifications
        push:
          type: boolean
          description: Enable push notifications
        frequency:
          type: string
          enum: ["immediate", "daily", "weekly"]
          description: Notification frequency
        types:
          type: array
          items:
            type: string
            enum: ["opportunity_match", "deadline_reminder", "roadmap_update"]
          description: Types of notifications to receive

    # AI Instructor Schemas
...
    Opportunity:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: ["hackathon", "internship", "workshop"]
        organizer:
          $ref: '#/components/schemas/Organizer'
        requirements:
          $ref: '#/components/schemas/Requirements'
        details:
          $ref: '#/components/schemas/OpportunityDetails'
        timeline:
          $ref: '#/components/schemas/Timeline'
        externalUrl:
          type: string
          format: uri
          description: Link to original opportunity posting
        sourceId:
          type: string
          description: ID of the data source
        tags:
          type: array
          items:
            type: string
          description: Opportunity tags
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        isActive:
          type: boolean
          description: Whether the opportunity is currently active
    Organizer:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: ["corporate", "startup", "government", "academic"]
        logo:
          type: string
          format: uri
          description: URL to organizer logo
    Requirements:
      type: object
      properties:
        skills:
          type: array
          items:
            type: string
          description: Required technical skills
        experience:
          type: string
          description: Required experience level
        education:
          type: string
          description: Educational requirements
        eligibility:
          type: array
          items:
            type: string
          description: Eligibility criteria
    OpportunityDetails:
      type: object
      properties:
        mode:
          type: string
          enum: ["online", "offline", "hybrid"]
        location:
          type: string
          description: Event location (for offline/hybrid events)
        duration:
          type: string
          description: Event duration
        stipend:
          type: string
          description: Stipend or compensation details
        prizes:
          type: array
          items:
            type: string
          description: Prize information
    Timeline:
      type: object
      properties:
        applicationDeadline:
          type: string
          format: date-time
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time

    # Notification Schemas
    Notification:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: ["opportunity_match", "deadline_reminder", "roadmap_update"]
        title:
          type: string
        message:
          type: string
        data:
          type: object
          description: Additional notification data
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # Common Schemas
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          default: 1
        limit:
          type: integer
          minimum: 1
          maximum: 50
          default: 20
        totalPages:
          type: integer
        totalCount:
          type: integer
    SearchQuery:
      type: object
      properties:
        query:
          type: string
        filters:
          $ref: '#/components/schemas/SearchFilters'
        timestamp:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: array
          items:
            type: string
          description: Additional error details
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/Error'
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "Invalid request parameters"
              details: ["Field 'query' is required"]
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/Error'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Invalid or expired token"
...
    RateLimit:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/Error'
          example:
            error:
              code: "RATE_LIMIT_EXCEEDED"
              message: "Too many requests. Please try again later."
...